requirements.txt
numpy>=1.24
pillow>=10.0
pyyaml>=6.0
tqdm>=4.66
scikit-learn>=1.3
albumentations>=1.4
scikit-image>=0.22
torch>=2.1
